chore: Restructure this repository as part of generalizing md2nb -> Markdown
md2nb is very specifically a command-line tool, but this same code
could be useful for more general processing of Markdown in Wolfram.

This commit is the first in a series that will 'rebrand' this repository
from md2nb to just Markdown, and add additional functionality like a
symbolic representation of Markdown in WL, and top-level WL functions
for creating notebooks from Markdown.

The old URL for this repository was:

	https://github.com/ConnorGray/md2nb

the new URL will be:

	https://github.com/ConnorGray/Markdown

The code at the old URL will be archived but for a small README.md change
pointing to the new repository.